Move theme related methods in GameUtils.java to new ThemeUtils.java utility class
